// frontend/src/types/generated/financial/models/costPoolResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Conecta PRO - Financial Module
 * Gestão Financeira Completa - 251 endpoints em 15 submódulos
 * OpenAPI spec version: 1.0.0
 */
import type { CostPoolResponseDescription } from './costPoolResponseDescription';
import type { PoolType } from './poolType';
import type { AllocationBasis } from './allocationBasis';
import type { CostPoolResponseParentId } from './costPoolResponseParentId';
import type { CostPoolResponseCostCenterId } from './costPoolResponseCostCenterId';
import type { CostPoolResponseCustomAllocationFormula } from './costPoolResponseCustomAllocationFormula';
import type { CostPoolResponseAllocationWeights } from './costPoolResponseAllocationWeights';
import type { CostPoolResponseManagerName } from './costPoolResponseManagerName';
import type { CostPoolResponseDepartment } from './costPoolResponseDepartment';
import type { CostPoolResponseNotes } from './costPoolResponseNotes';
import type { CostPoolResponseTags } from './costPoolResponseTags';
import type { PoolStatus } from './poolStatus';
import type { CostPoolResponsePath } from './costPoolResponsePath';

/**
 * Schema de resposta para Cost Pool.
 */
export interface CostPoolResponse {
  /**
   * @minLength 1
   * @maxLength 30
   */
  code: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  description?: CostPoolResponseDescription;
  pool_type?: PoolType;
  allocation_basis?: AllocationBasis;
  parent_id?: CostPoolResponseParentId;
  cost_center_id?: CostPoolResponseCostCenterId;
  budget_amount?: string;
  custom_allocation_formula?: CostPoolResponseCustomAllocationFormula;
  allocation_weights?: CostPoolResponseAllocationWeights;
  manager_name?: CostPoolResponseManagerName;
  department?: CostPoolResponseDepartment;
  is_homogeneous?: boolean;
  auto_allocate?: boolean;
  notes?: CostPoolResponseNotes;
  tags?: CostPoolResponseTags;
  id: string;
  condominio_id: string;
  status: PoolStatus;
  level: number;
  path: CostPoolResponsePath;
  total_cost: string;
  allocated_cost: string;
  unallocated_cost: string;
  budget_variance: string;
  allocation_base_quantity: string;
  allocation_rate: string;
  activities_count: number;
  allocations_count: number;
  active: boolean;
  created_at: string;
  updated_at: string;
}